refactor(callback): replace worker
